+++ lib/src/embedded/importer/host.dart
@@ -6,7 +6,6 @@ import '../../exception.dart';
 import '../../importer.dart';
 import '../../importer/utils.dart';
 import '../../util/span.dart';
-import '../compilation_dispatcher.dart';
 import '../embedded_sass.pb.dart' hide SourceSpan;
 import '../utils.dart';
 import 'base.dart';
@@ -20,10 +19,9 @@ final class HostImporter extends ImporterBase {
   /// [canonicalize].
   final Set<String> _nonCanonicalSchemes;
 
-  HostImporter(CompilationDispatcher dispatcher, this._importerId,
-      Iterable<String> nonCanonicalSchemes)
-      : _nonCanonicalSchemes = Set.unmodifiable(nonCanonicalSchemes),
-        super(dispatcher) {
+  HostImporter(
+      super.dispatcher, this._importerId, Iterable<String> nonCanonicalSchemes)
+      : _nonCanonicalSchemes = Set.unmodifiable(nonCanonicalSchemes) {
     for (var scheme in _nonCanonicalSchemes) {
       if (isValidUrlScheme(scheme)) continue;
       throw SassException(