+++ lib/src/parse/media_query.dart
@@ -5,17 +5,13 @@
 import 'package:charcode/charcode.dart';
 
 import '../ast/css.dart';
-import '../interpolation_map.dart';
-import '../logger.dart';
 import '../utils.dart';
 import 'parser.dart';
 
 /// A parser for `@media` queries.
 class MediaQueryParser extends Parser {
-  MediaQueryParser(String contents,
-      {Object? url, Logger? logger, InterpolationMap? interpolationMap})
-      : super(contents,
-            url: url, logger: logger, interpolationMap: interpolationMap);
+  MediaQueryParser(super.contents,
+      {super.url, super.logger, super.interpolationMap});
 
   List<CssMediaQuery> parse() {
     return wrapSpanFormatException(() {