+++ tool/grind.dart
@@ -29,11 +29,14 @@ void main(List<String> args) {
   pkg.chocolateyNuspec.value = _nuspec;
   pkg.homebrewRepo.value = "sass/homebrew-sass";
   pkg.homebrewFormula.value = "Formula/sass.rb";
+  pkg.homebrewEditFormula.value = _updateHomebrewLanguageRevision;
   pkg.jsRequires.value = [
     pkg.JSRequire("immutable", target: pkg.JSRequireTarget.all),
     pkg.JSRequire("chokidar", target: pkg.JSRequireTarget.cli),
     pkg.JSRequire("readline", target: pkg.JSRequireTarget.cli),
     pkg.JSRequire("fs", target: pkg.JSRequireTarget.node),
+    pkg.JSRequire("module",
+        target: pkg.JSRequireTarget.node, identifier: 'nodeModule'),
     pkg.JSRequire("stream", target: pkg.JSRequireTarget.node),
     pkg.JSRequire("util", target: pkg.JSRequireTarget.node),
   ];
@@ -51,6 +54,10 @@ void main(List<String> args) {
     'compileAsync',
     'compileString',
     'compileStringAsync',
+    'initCompiler',
+    'initAsyncCompiler',
+    'Compiler',
+    'AsyncCompiler',
     'Logger',
     'SassArgumentList',
     'SassBoolean',
@@ -79,6 +86,7 @@ void main(List<String> args) {
     'FALSE',
     'NULL',
     'types',
+    'NodePackageImporter',
   };
 
   pkg.githubReleaseNotes.fn = () =>
@@ -287,4 +295,31 @@ function defaultExportDeprecation() {
   buffer.writeln("};");
 
   File("build/npm/sass.node.mjs").writeAsStringSync(buffer.toString());
+}
+
+/// A regular expression to locate the language repo revision in the Dart Sass
+/// Homebrew formula.
+final _homebrewLanguageRegExp = RegExp(
+    r'resource "language" do$'
+    r'(?:(?! end$).)+'
+    r'revision: "([a-f0-9]{40})"',
+    dotAll: true,
+    multiLine: true);
+
+/// Updates the Homebrew [formula] to change the revision of the language repo
+/// to the latest revision.
+String _updateHomebrewLanguageRevision(String formula) {
+  var languageRepoRevision = run("git",
+          arguments: ["ls-remote", "https://github.com/sass/sass"], quiet: true)
+      .split("\t")
+      .first;
+
+  var match = _homebrewLanguageRegExp.firstMatch(formula);
+  if (match == null) {
+    fail("Couldn't find a language repo revision in the Homebrew formula.");
+  }
+
+  return formula.substring(0, match.start) +
+      match.group(0)!.replaceFirst(match.group(1)!, languageRepoRevision) +
+      formula.substring(match.end);
 }